Module phar::read::index

This is supported on crate feature reader only.

FileIndex implementations

Structs

MetadataMap

Indexes files by name for random access, and stores file metadata.

NameMap

Indexes files by name for random access.

NoIndex

Indicates that the phar should not index phar files at all.

OffsetOnly

Stores files by position.

Traits

FileIndex

The storage used to store file indices.

Iterable

A subfamily of file indices for iterable files.

RandomAccess

A subfamily of file indices for random access of files by name.

Type Definitions

MetadataBTreeMap

Indexes files by name with a BTreeMap, and stores file metadata.

MetadataHashMap

Indexes files by name with a HashMap, and stores file metadata.

NameBTreeMap

Indexes files by name with a BTreeMap.

NameHashMap

Indexes files by name with a HashMap.