Trait phar::read::index::FileIndex

pub trait FileIndex: Default {
    fn feed_entry(&mut self, offset: u64, entry: Entry) -> Result<()>;

    fn scan_files() -> bool { ... }

    fn requires_name() -> bool { ... }

    fn requires_metadata() -> bool { ... }

    fn end_of_header(&mut self, _offset: u64) { ... }
}

This is supported on crate feature reader only.

The storage used to store file indices.

If you open the phar only to view the stub, phar metadata, etc., use index::NoIndex.

To sequentially access all files in the archive, use index::Iterable implementations. If random file access is not required, use index::OffsetOnly.

To access specific files only, use index::RandomAccess implementations. Prefer using NameMap if individual entry metadata is not required. To also access their metadata, use MetadataMap. There are some type aliases for the respective HashMap/BTreeMap implementations.

Required methods

fn feed_entry(&mut self, offset: u64, entry: Entry) -> Result<()>

Adds an Entry to the index.

Provided methods

fn scan_files() -> bool

Whether file metadata should be scanned on loading.

fn requires_name() -> bool

Whether Entry should force name to use Section::Cached.

fn requires_metadata() -> bool

Whether Entry should force metadata to use Section::Cached.

fn end_of_header(&mut self, _offset: u64)

Marks the end of header