Trait phar::read::index::Iterable

pub trait Iterable: FileIndex {
    fn for_each_file_fold<'t, R, F, G, T, U>(
        &self, 
        read: R, 
        f: F, 
        fold: G
    ) -> Result<Option<T>>
    where
        R: Read + Seek + 't,
        F: FnMut(&[u8], &mut dyn Read) -> Result<U>,
        G: FnMut(Option<T>, U) -> T;

    fn for_each_file<'t, R, F>(&self, read: R, f: F) -> Result<()>
    where
        R: Read + Seek + 't,
        F: FnMut(&[u8], &mut dyn Read) -> Result<()>,
    { ... }
}

This is supported on crate feature reader only.

A subfamily of file indices for iterable files.

The iteration order may not be the order in the phar archive, and may not even be stable.

Required methods

fn for_each_file_fold<'t, R, F, G, T, U>(
&self,
read: R,
f: F,
fold: G
) -> Result<Option<T>> where
R: Read + Seek + 't,
F: FnMut(&[u8], &mut dyn Read) -> Result<U>,
G: FnMut(Option<T>, U) -> T,

Iterates over the files in this index and fold return values.

Provided methods

fn for_each_file<'t, R, F>(&self, read: R, f: F) -> Result<()> where
R: Read + Seek + 't,
F: FnMut(&[u8], &mut dyn Read) -> Result<()>,

Iterates over the files in this index.