Trait phar::read::index::RandomAccess

pub trait RandomAccess: FileIndex {
    fn read_file(&self, name: &[u8]) -> Option<Range<u64>>;
}

This is supported on crate feature reader only.

A subfamily of file indices for random access of files by name.

Required methods

fn read_file(&self, name: &[u8]) -> Option<Range<u64>>

Returns the file contents range of the file of the required name.

Returns None if there are no files with the specified name.

Loading content...