Struct phar::Reader

pub struct Reader<R: Read + Seek, FileIndexT: FileIndex = NameHashMap> { /* fields omitted */ }

This is supported on crate feature reader only.

The metadata of a phar file.

Implementations

impl<R: Read + Seek, FileIndexT: FileIndex> Reader<R, FileIndexT>

pub fn read(read: R, options: Options) -> Result<Self>

Parses the phar file.

This optionally validates the signature. Stub, metadata and file metadata are not fully validated, and may not be saved in memory depending on options.

pub fn stub_bytes(&mut self) -> Result<impl AsRef<[u8]> + '_>

Returns the stub as a slice.

If the stub was previously not stored in memory, it is stored in a new Vec. Consider using stub_read() instead if cache_stub is false and storing the stub in memory is not intended.

pub fn stub_read(&mut self) -> Result<impl Read + '_>

Returns the stub as an io::Read.

pub fn metadata_bytes(&mut self) -> Result<impl AsRef<[u8]> + '_>

Returns the metadata as a slice.

If the metadata was previously not stored in memory, it is stored in a new Vec. Consider using metadata_read() instead if cache_metadata is false and storing the metadata in memory is not intended.

pub fn metadata_read(&mut self) -> Result<impl Read + '_>

Returns the metadata as an io::Read.

impl<R: Read + Seek, FileIndexT: Iterable> Reader<R, FileIndexT>

pub fn for_each_file<F>(&mut self, f: F) -> Result<()> where
F: FnMut(&[u8], &mut dyn Read) -> Result<()>,

Iterates over the files in this archive.

pub fn for_each_file_fold<F, G, T, U>(
&mut self,
f: F,
fold: G
) -> Result<Option<T>> where
F: FnMut(&[u8], &mut dyn Read) -> Result<U>,
G: FnMut(Option<T>, U) -> T,

Iterates over the files in this archive and fold return values.